refactoring edit post and fixing delete bug
