+++ data/db.json
@@ -1,22 +1,22 @@
 {
   "posts": [
     {
-      "id": 1,
-      "title": "1st post",
-      "datetime": "July 21, 2022 3:45:45 PM",
-      "body": "full of shit "
+      "id": 6,
+      "title": "test",
+      "datetime": "July 21, 2022 8:47:35 PM",
+      "body": "asdsfa"
     },
     {
-      "id": 2,
-      "title": "dave gray ",
-      "datetime": "July 21, 2022 12:57:20 PM",
-      "body": "this is modified comment by Sam Taklimi "
+      "id": 7,
+      "title": "test4",
+      "datetime": "July 21, 2022 11:25:32 PM",
+      "body": "ading 4 statement"
     },
     {
-      "id": 3,
-      "title": "Saeed Njaf",
-      "datetime": "July 21, 2022 5:03:47 PM",
-      "body": "this is also modified by me as i hate lerom text "
+      "id": 8,
+      "title": "test2",
+      "datetime": "July 21, 2022 11:25:48 PM",
+      "body": "this id test 2 "
     }
   ]
 }
