+++ src/EditPost.js
@@ -1,12 +1,17 @@
 import React from "react";
-import { useEffect, useContext } from "react";
-import { useParams, Link } from "react-router-dom";
+import { useEffect, useContext, useState } from "react";
+import { useParams, Link, useNavigate } from "react-router-dom";
 import DataContext from "./context/DataContext";
+import api from "./api/posts";
+import { format } from "date-fns";
 
 const EditPost = () => {
-  const { posts, handleEdit, editBody, setEditBody, editTitle, setEditTitle } = useContext(DataContext);
+  const [editTitle, setEditTitle] = useState("");
+  const [editBody, setEditBody] = useState("");
+  const { posts, setPosts } = useContext(DataContext);
   const { id } = useParams();
   const post = posts.find((post) => post.id.toString() === id);
+  const navigate = useNavigate();
 
   useEffect(() => {
     if (post) {
@@ -14,6 +19,22 @@ const EditPost = () => {
       setEditBody(post.body);
     }
   }, [post, setEditTitle, setEditBody]);
+  const handleEdit = async (id) => {
+    const datetime = format(new Date(), "MMMM dd, yyyy pp");
+    const updatedPost = { id, title: editTitle, datetime, body: editBody };
+    try {
+      const response = await api.put(`/posts/${id}`, updatedPost);
+      setPosts(
+        posts.map((post) => (post.id === id ? { ...response.data } : post))
+      );
+      setEditTitle("");
+      setEditBody("");
+      navigate("/");
+    } catch (err) {
+      console.log(`Error: ${err.message}`);
+    }
+  };
+
   return (
     <main className="NewPost">
       {editTitle && (
@@ -35,20 +56,21 @@ const EditPost = () => {
               value={editBody}
               onChange={(e) => setEditBody(e.target.value)}
             />
-            <button type="submit" onClick={()=> handleEdit(post.id)}>Submit</button>
+            <button type="submit" onClick={() => handleEdit(post.id)}>
+              Submit
+            </button>
           </form>
         </>
       )}
-      {!editTitle && 
-          <>
-            <h2>Post Not Found</h2>
-            <p>Well, that's disappointing.</p>
-            <p>
-              <Link to='/'>Vist Our Homepage</Link>
-            </p>
-          </>
-
-        }
+      {!editTitle && (
+        <>
+          <h2>Post Not Found</h2>
+          <p>Well, that's disappointing.</p>
+          <p>
+            <Link to="/">Vist Our Homepage</Link>
+          </p>
+        </>
+      )}
     </main>
   );
 };
