+++ src/PostPage.js
@@ -1,12 +1,26 @@
 import React from "react";
-import { useParams, Link } from "react-router-dom";
+import { useParams, Link ,useNavigate } from "react-router-dom";
 import { useContext } from "react";
 import DataContext from "./context/DataContext";
+import api from './api/posts';
 
 const PostPage = () => {
-  const { posts, handleDelete } = useContext(DataContext);
+  const { posts,setPosts } = useContext(DataContext);
   const { id } = useParams();
   const post = posts.find((post) => post.id.toString() === id);
+  const navigate = useNavigate();
+
+
+  const handleDelete = async (id) => {
+    try {
+      await api.delete(`/posts/${id}`);
+      const postsList = posts.filter((post) => post.id !== id);
+      setPosts(postsList);
+      navigate("/");
+    } catch (err) {
+      console.log(`Error2: ${err.message}`);
+    }
+  };
   return (
     <main className="PostPage">
       <article className="post">
