+++ src/context/DataContext.js
@@ -1,6 +1,5 @@
 import { useEffect, useState, createContext } from "react";
-import { Route, Routes, useNavigate } from "react-router-dom";
-import { format } from "date-fns";
+import { useNavigate } from "react-router-dom";
 import api from "../api/posts";
 import useAxiosFetch from "../hooks/useAxiosFetch";
 
@@ -11,8 +10,7 @@ export const DataProvider = ({ children }) => {
   const [search, setSearch] = useState("");
   const [searchResults, setSearchResults] = useState([]);
  
-  const [editTitle, setEditTitle] = useState("");
-  const [editBody, setEditBody] = useState("");
+  
   const navigate = useNavigate();
   const { data, fetchError, isLoading } = useAxiosFetch(
     "http://localhost:3500/posts"
@@ -49,34 +47,7 @@ export const DataProvider = ({ children }) => {
     );
     setSearchResults(filteredResults.reverse());
   }, [posts, search]);
-
-  
-
-  const handleEdit = async (id) => {
-    const datetime = format(new Date(), "MMMM dd, yyyy pp");
-    const updatedPost = { id, title: editTitle, datetime, body: editBody };
-    try {
-      const response = await api.put(`/posts/${id}`, updatedPost);
-      setPosts(
-        posts.map((post) => (post.id === id ? { ...response.data } : post))
-      );
-      setEditTitle("");
-      setEditBody("");
-      navigate("/");
-    } catch (err) {
-      console.log(`Error: ${err.message}`);
-    }
-  };
-  const handleDelete = async (id) => {
-    try {
-      await api.delete(`/posts/${id}`);
-      const postsList = posts.filter((post) => post.id !== id);
-      setPosts(postsList);
-      navigate("/");
-    } catch (err) {
-      console.log(`Error: ${err.message}`);
-    }
-  };
+ 
   return (
     <DataContext.Provider
       value={{
@@ -86,12 +57,7 @@ export const DataProvider = ({ children }) => {
         fetchError,
         isLoading,
         posts,
-        handleEdit,
-        editBody,
-        setEditBody,
-        editTitle,
-        setEditTitle,
-        handleDelete,
+        setPosts,
       }}
     >
       {children}
